refactor: decouple routingServer into BGPWatcher, BGPHandler and routingHandler
The refactoring splits the monolithic routingServer into three focused components:
i) BGPWatcher - observes GoBGP RIB state
ii) BGPHandler - handles BGP protocol business logic
iii) RoutingHandler - handles route installation business logic

There is a clear separation of monitoring (BGPWatcher) and business logic:
i) GoBGP programming via the BGPHandler
ii) Linux Kernel programming via the routingHandler
iii) VPP programming via the connectivityHandler
