+++ calico-vpp-agent/felix/routing/peer_handler.go
@@ -28,7 +28,6 @@ import (
 
 	"github.com/projectcalico/vpp-dataplane/v3/calico-vpp-agent/common"
 	"github.com/projectcalico/vpp-dataplane/v3/calico-vpp-agent/felix/cache"
-	"github.com/projectcalico/vpp-dataplane/v3/calico-vpp-agent/watchers"
 	"github.com/projectcalico/vpp-dataplane/v3/config"
 )
 
@@ -252,7 +251,7 @@ func (h *PeerHandler) addBGPPeer(ip string, asn uint32, peerSpec *calicov3.BGPPe
 	}
 	common.SendEvent(common.CalicoVppEvent{
 		Type: common.BGPPeerAdded,
-		New:  &watchers.LocalBGPPeer{Peer: peer, BGPFilterNames: peerSpec.Filters},
+		New:  &LocalBGPPeer{Peer: peer, BGPFilterNames: peerSpec.Filters},
 	})
 	return nil
 }
@@ -264,8 +263,8 @@ func (h *PeerHandler) updateBGPPeer(ip string, asn uint32, peerSpec, oldPeerSpec
 	}
 	common.SendEvent(common.CalicoVppEvent{
 		Type: common.BGPPeerUpdated,
-		New:  &watchers.LocalBGPPeer{Peer: peer, BGPFilterNames: peerSpec.Filters},
-		Old:  &watchers.LocalBGPPeer{BGPFilterNames: oldPeerSpec.Filters},
+		New:  &LocalBGPPeer{Peer: peer, BGPFilterNames: peerSpec.Filters},
+		Old:  &LocalBGPPeer{BGPFilterNames: oldPeerSpec.Filters},
 	})
 	return nil
 }
@@ -273,7 +272,7 @@ func (h *PeerHandler) updateBGPPeer(ip string, asn uint32, peerSpec, oldPeerSpec
 func (h *PeerHandler) deleteBGPPeer(ip string) error {
 	common.SendEvent(common.CalicoVppEvent{
 		Type: common.BGPPeerDeleted,
-		New:  ip,
+		Old:  ip,
 	})
 	return nil
 }
