+++ calico-vpp-agent/watchers/peers_watcher.go
@@ -17,11 +17,8 @@ package watchers
 
 import (
 	"context"
-	"reflect"
-	"sort"
 	"time"
 
-	bgpapi "github.com/osrg/gobgp/v3/api"
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
 	"gopkg.in/tomb.v2"
@@ -35,24 +32,6 @@ import (
 
 	"github.com/projectcalico/vpp-dataplane/v3/calico-vpp-agent/common"
 )
-
-type LocalBGPPeer struct {
-	Peer           *bgpapi.Peer
-	BGPFilterNames []string
-	BGPPolicies    map[string]*ImpExpPol
-	NeighborSet    *bgpapi.DefinedSet
-}
-
-type BGPPrefixesPolicyAndAssignment struct {
-	PolicyAssignment *bgpapi.PolicyAssignment
-	Policy           *bgpapi.Policy
-	Prefixes         []*bgpapi.DefinedSet
-}
-
-type ImpExpPol struct {
-	Imp *BGPPrefixesPolicyAndAssignment
-	Exp *BGPPrefixesPolicyAndAssignment
-}
 
 type PeerWatcher struct {
 	log      *logrus.Entry
@@ -156,19 +135,6 @@ func (w *PeerWatcher) WatchBGPPeers(t *tomb.Tomb) error {
 	}
 	w.log.Warn("BGPPeer watcher stopped")
 	return nil
-}
-
-func CompareStringSlices(slice1, slice2 []string) bool {
-	if len(slice1) != len(slice2) {
-		return false
-	}
-
-	// Sort the slices in ascending order
-	sort.Strings(slice1)
-	sort.Strings(slice2)
-
-	// Compare the sorted slices
-	return reflect.DeepEqual(slice1, slice2)
 }
 
 func (w *PeerWatcher) resyncAndCreateWatcher() error {
